+++ GetIn/Screens/GIWordsVC.swift
@@ -190,26 +190,38 @@ extension GIWordsVC : UITableViewDelegate, UITableViewDataSource {
     func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
         let action = UIContextualAction(style: .destructive, title: "Delete") { [weak self] ( _ , view, completed) in
             
-            guard let vc = self else { return }
-            
-            guard let wordToRemove = vc.words?[indexPath.row] else { return }
-            vc.container?.viewContext.delete(wordToRemove)
-            
-            do {
-                try self?.container?.viewContext.save()
-            } catch {
-                print("save when delete error")
+            let ac = UIAlertController(title: "Delete?", message: "Do you really want to delete this word? You won't be able to restore it.", preferredStyle: .alert)
+            let confirmAction = UIAlertAction(title: "Delete", style: .destructive) { (action) in
+                
+                guard let vc = self else { return }
+                
+                guard let wordToRemove = vc.words?[indexPath.row] else { return }
+                vc.container?.viewContext.delete(wordToRemove)
+                
+                do {
+                    try self?.container?.viewContext.save()
+                } catch {
+                    print("save when delete error")
+                }
+                
+                vc.words?.remove(at: indexPath.row)
+                
+                tableView.deleteRows(at: [indexPath], with: .fade)
+                completed(true)
+                
+                if self!.words?.count == 0 {
+                    self!.configureEmptyStateView(with: "No words here.\nAdd some :)", in: view)
+                    self?.tableView.isHidden = true
+                }
+                
             }
-            
-            vc.words?.remove(at: indexPath.row)
-            
-            tableView.deleteRows(at: [indexPath], with: .fade)
-            completed(true)
-            
-            if self!.words?.count == 0 {
-                self!.configureEmptyStateView(with: "No words here.\nAdd some :)", in: view)
-                self?.tableView.isHidden = true
+            let cancelAction = UIAlertAction(title: "Cancel", style: .cancel) { (action) in
+                self?.tableView.reloadRows(at: [indexPath], with: .automatic)
             }
+            ac.addAction(confirmAction)
+            ac.addAction(cancelAction)
+            
+            self?.present(ac, animated: true)
         }
         action.backgroundColor = .systemRed
         return UISwipeActionsConfiguration(actions: [action])
